requirements.txt
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.5.0
langchain-community==0.2.0
python-multipart==0.0.6
openai==1.0.0  # For OpenRouter compat
